+++ functions/src/actions/fetchChannels.ts
@@ -1,4 +1,5 @@
 import * as admin from 'firebase-admin';
+import { FIXME } from '../../custom';
 
 interface Channel {
   key: string;
