+++ functions/src/actions/index.ts
@@ -1,0 +1,5 @@
+import { fetchChannels } from './fetchChannels';
+import { createChannel } from './createChannel';
+import { createMessage } from './createMessage';
+
+export { createChannel, createMessage, fetchChannels };
